Add structured Asset Lab reject reasons with separate manual evidence.
Manual review can mark validator misses or false detections, and those human reasons stay on the selected asset instead of leaking through a global Berwynn list.

+++ server/learning/policy.py
@@ -38,7 +38,7 @@ def unit_interval(value: float) -> float:
     return max(0.0, min(1.0, float(value)))
 
 
-def confidence_for(evidence: int, consistency: float, policy: LearningPolicy) -> tuple[float, bool, str]:
+def confidence_for(evidence: float, consistency: float, policy: LearningPolicy) -> tuple[float, bool, str]:
     """confidence = unit(band_scale) * unit(consistency). Always in [0, 1]."""
     consistency = unit_interval(consistency)
     if evidence < policy.minSuggest:
